requirements.txt
apache-airflow
streamlit
langchain
sentence-transformers
pypdf
faiss-cpu
openai
langchain-community
tiktoken